refactor(review): 도메인 root flat 구조를 sub-folder로 분류
- root 18개 파일을 blog/, modal/, form/, ui/, section/ 폴더로 이동
- ReviewCard 322줄 → 183줄로 축소 (ReviewItemBlock, reviewCardHelpers 추출)
- regacy → legacy 오타 수정 (4파일 + 외부 import 1곳)
- 상대경로 (../../common/...) → @/ alias로 통일
