+++ apps/web/src/app/(user)/dev/page.tsx
@@ -2,7 +2,7 @@ import { GetReview } from '@/api/review/review';
 import FilterDropdown from '@/common/dropdown/FilterDropdown';
 import MoreHeader from '@/common/header/MoreHeader';
 import ReviewSection from '@/domain/about/section/ReviewSection';
-import ReviewCard from '@/domain/review/ReviewCard';
+import ReviewCard from '@/domain/review/ui/ReviewCard';
 import { Metadata } from 'next';
 import Image from 'next/image';
 import { Suspense } from 'react';
