+++ apps/web/src/app/(user)/mypage/review/report/[reportId]/page.tsx
@@ -10,11 +10,11 @@ import { Suspense } from 'react';
 import { useGetReportTitle } from '@/api/report';
 import { useGetProgramReviewDetail } from '@/api/review/review';
 import { useUserQuery } from '@/api/user/user';
-import GoalOrConcernsBox from '@/domain/review/GoalOrConcernsBox';
-import ReviewInstruction from '@/domain/review/ReviewInstruction';
-import ReviewModal from '@/domain/review/ReviewModal';
-import ReviewQuestion from '@/domain/review/ReviewQuestion';
-import ReviewTextarea from '@/domain/review/ReviewTextarea';
+import GoalOrConcernsBox from '@/domain/review/form/GoalOrConcernsBox';
+import ReviewInstruction from '@/domain/review/form/ReviewInstruction';
+import ReviewModal from '@/domain/review/modal/ReviewModal';
+import ReviewQuestion from '@/domain/review/form/ReviewQuestion';
+import ReviewTextarea from '@/domain/review/form/ReviewTextarea';
 import TenScore from '@/domain/review/score/TenScore';
 
 const ReportReviewPageContent = () => {
