+++ apps/web/src/domain/challenge/my-challenge/section/LastMissionSubmitModal.tsx
@@ -4,11 +4,11 @@ import {
 } from '@/api/challenge/challenge';
 import { usePostReviewMutation } from '@/api/review/review';
 import { useUserQuery } from '@/api/user/user';
-import GoalOrConcernsBox from '@/domain/review/GoalOrConcernsBox';
-import ReviewInstruction from '@/domain/review/ReviewInstruction';
-import ReviewModal from '@/domain/review/ReviewModal';
-import ReviewQuestion from '@/domain/review/ReviewQuestion';
-import ReviewTextarea from '@/domain/review/ReviewTextarea';
+import GoalOrConcernsBox from '@/domain/review/form/GoalOrConcernsBox';
+import ReviewInstruction from '@/domain/review/form/ReviewInstruction';
+import ReviewModal from '@/domain/review/modal/ReviewModal';
+import ReviewQuestion from '@/domain/review/form/ReviewQuestion';
+import ReviewTextarea from '@/domain/review/form/ReviewTextarea';
 import TenScore from '@/domain/review/score/TenScore';
 import { useMediaQuery } from '@mui/material';
 import { josa } from 'es-hangul';
