+++ apps/web/src/domain/review/blog/MainBlogReviewSection.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { useGetBlogReviewList } from '@/api/review/review';
-import MoreHeader from '../../common/header/MoreHeader';
-import LoadingContainer from '../../common/loading/LoadingContainer';
-import ReviewLinkCard from './ReviewLinkCard';
+import MoreHeader from '@/common/header/MoreHeader';
+import LoadingContainer from '@/common/loading/LoadingContainer';
+import ReviewLinkCard from '../ReviewLinkCard';
 
 function MainBlogReviewSection() {
   const { data, isLoading } = useGetBlogReviewList({
