+++ apps/web/src/domain/review/legacy/ui/Star.tsx
