+++ apps/web/src/domain/review/programReview/ProgramReviewSection.tsx
@@ -3,7 +3,7 @@
 import { GetReview, useGetProgramReview } from '@/api/review/review';
 import MoreHeader from '@/common/header/MoreHeader';
 import LoadingContainer from '@/common/loading/LoadingContainer';
-import ReviewCard from '@/domain/review/ReviewCard';
+import ReviewCard from '@/domain/review/ui/ReviewCard';
 import { useEffect, useState } from 'react';
 
 const ProgramReviewSection = () => {
