+++ apps/web/src/domain/review/ui/ReviewCard.tsx
@@ -1,106 +1,19 @@
 'use client';
 
-import { ReportType } from '@/api/report';
-import { GetReview, QuestionType, ReviewType } from '@/api/review/review';
-import Bubble from '@/assets/graphic/bubble.svg?react';
-import Heart from '@/assets/graphic/heart.svg?react';
-import Lightbulb from '@/assets/graphic/lightbulb.svg?react';
-import Pen from '@/assets/graphic/pen.svg?react';
-import PinBlue from '@/assets/graphic/pin_blue.svg?react';
-import PinRed from '@/assets/graphic/pin_red.svg?react';
-import Trophy from '@/assets/graphic/trophy.svg?react';
+import { GetReview } from '@/api/review/review';
 import { YYYY_MM_DD } from '@/data/dayjsFormat';
 import dayjs from '@/lib/dayjs';
 import { twMerge } from '@/lib/twMerge';
-import { ProgramTypeUpperCase } from '@/schema';
-import { questionTypeToText } from '@/utils/convert';
 import clsx from 'clsx';
 import { useRouter } from 'next/navigation';
-import { ReactNode } from 'react';
-import ExpandableParagraph from './ExpandableParagraph';
 import ReviewBadge from './ReviewBadge';
-
-function reviewTypeToProgramType(reviewType: ReviewType): ProgramTypeUpperCase {
-  switch (reviewType) {
-    case 'CHALLENGE_REVIEW':
-    case 'MISSION_REVIEW':
-      return 'CHALLENGE';
-    case 'LIVE_REVIEW':
-      return 'LIVE';
-    case 'VOD_REVIEW':
-      return 'VOD';
-    case 'REPORT_REVIEW':
-      return 'REPORT';
-  }
-}
-
-export const getTitle = (review: GetReview) => {
-  switch (review.reviewInfo.type) {
-    case 'CHALLENGE_REVIEW':
-    case 'LIVE_REVIEW':
-    case 'VOD_REVIEW':
-      return review.reviewInfo.programTitle;
-    case 'REPORT_REVIEW':
-      // TODO: 서류 진단 제목이 이 필드가 맞는지 체크 필요.
-      return review.reviewInfo.programTitle;
-    case 'MISSION_REVIEW':
-      return review.reviewInfo.missionTitle;
-  }
-};
-
-const questionIcon = (questionType: QuestionType | null) => {
-  switch (questionType) {
-    case 'WORRY':
-      return <PinBlue width={18} height={18} />;
-    case 'WORRY_RESULT':
-      return <Lightbulb width={18} height={18} />;
-    case 'GOAL':
-      return <PinRed width={18} height={18} />;
-    case 'GOAL_RESULT':
-      return <Trophy width={18} height={18} />;
-    case 'GOOD_POINT':
-      return <Heart width={18} height={18} />;
-    case 'BAD_POINT':
-      return <Pen width={18} height={18} />;
-    default:
-      return <Bubble width={18} height={18} />;
-  }
-};
-
-const questionPriority = (questionType: QuestionType | null) => {
-  switch (questionType) {
-    case 'GOAL':
-      return 1;
-    case 'GOAL_RESULT':
-      return 2;
-    case 'WORRY':
-      return 3;
-    case 'WORRY_RESULT':
-      return 4;
-    case 'GOOD_POINT':
-      return 5;
-    case 'BAD_POINT':
-      return 6;
-    default:
-      return 7;
-  }
-};
-
-const getThumbnail = (
-  reportType: ReportType | null,
-  programThumbnail: string | null,
-) => {
-  switch (reportType) {
-    case 'RESUME':
-      return '/images/report/thumbnail_resume.png';
-    case 'PERSONAL_STATEMENT':
-      return '/images/report/thumbnail_personal.png';
-    case 'PORTFOLIO':
-      return '/images/report/thumbnail_portfolio.png';
-    default:
-      return programThumbnail ?? '';
-  }
-};
+import ReviewItemBlock from './ReviewItemBlock';
+import {
+  getThumbnail,
+  questionIcon,
+  questionPriority,
+  reviewTypeToProgramType,
+} from './reviewCardHelpers';
 
 const ReviewCard = ({
   review,
@@ -263,58 +176,6 @@ const ReviewCard = ({
           }}
         />
       ) : null}
-    </div>
-  );
-};
-
-const ReviewItemBlock = (props: {
-  answer?: string | null;
-  questionType?: QuestionType | null;
-  lineClamp?: 1 | 2 | 3 | 4 | 5;
-  /** 이게 있을 경우 우선함. */
-  questionText?: string;
-  expandable?: boolean;
-  icon?: ReactNode;
-  gap?: 'normal' | 'large';
-}) => {
-  const questionText =
-    props.questionText ||
-    (props.questionType ? questionTypeToText[props.questionType] : '');
-
-  return (
-    <div>
-      <div className="flex w-fit items-center gap-1">
-        {/* {props.icon && props.icon} */}
-        <span className="text-xsmall14 text-neutral-10 font-semibold">
-          {questionText}
-        </span>
-      </div>
-      {props.expandable ? (
-        <ExpandableParagraph
-          content={props.answer ?? ''}
-          lineClamp={props.lineClamp}
-          className={twMerge('text-xsmall14 text-neutral-20 font-normal')}
-        />
-      ) : (
-        <p
-          className={twMerge(
-            'text-xsmall14 text-neutral-20 font-normal tracking-[-0.028px]',
-            props.lineClamp === 1
-              ? 'line-clamp-1'
-              : props.lineClamp === 2
-                ? 'line-clamp-2'
-                : props.lineClamp === 3
-                  ? 'line-clamp-3'
-                  : props.lineClamp === 4
-                    ? 'line-clamp-4'
-                    : props.lineClamp === 5
-                      ? 'line-clamp-5'
-                      : null,
-          )}
-        >
-          {props.answer}
-        </p>
-      )}
     </div>
   );
 };
