+++ apps/web/src/domain/review/ui/ReviewLinkCard.tsx
@@ -3,7 +3,7 @@
 import { YYYY_MM_DD } from '@/data/dayjsFormat';
 import ReviewBadge, {
   getBadgeTypeFromProgramType,
-} from '@/domain/review/ReviewBadge';
+} from '@/domain/review/ui/ReviewBadge';
 import dayjs from '@/lib/dayjs';
 import { twMerge } from '@/lib/twMerge';
 import { ProgramTypeUpperCase } from '@/schema';
